FIXED GYRO

+++ Assignment_5/graphicsLab/project/data/TAL.vert
@@ -1,7 +1,7 @@
 
-uniform mediump mat4 ViewMatrix;
-uniform mediump mat4 modelMatrixTAL;
-uniform mediump mat4 ProjectionMatrix;
+uniform highp mat4 ViewMatrix;
+uniform highp mat4 modelMatrixTAL;
+uniform highp mat4 ProjectionMatrix;
 
 uniform mediump mat3 NormalMatrixTAL;
